+++ src/main/java/lotto/LottoMachine.java
@@ -1,41 +1,17 @@
 package lotto;
 
-import lotto.domain.Lotto;
-import lotto.domain.LottoAmount;
-import lotto.domain.LottoNumberStrategy;
-import lotto.domain.PurchasedLotto;
-
-import java.util.*;
+import lotto.view.LottoResultView;
 
 /**
  * @author : 김윤호
  * @version : 1.0
- * @date : 2019-11-02 22:27
+ * @date : 2019-11-02 22:36
  */
 public class LottoMachine {
+    public static void main(String[] args) {
+        LottoOrder lottoMachine = new LottoOrder(14000);
+        LottoResultView.printLotto(lottoMachine.getLottoNumbers());
 
-    private static final LottoNumberStrategy LOTTO_NUMBER_STRATEGY = new LottoNumberStrategy();
-    private final LottoAmount lottoAmount;
-    private final PurchasedLotto purchasedLotto;
 
-    public LottoMachine(int amount) {
-        this.lottoAmount = new LottoAmount(amount);
-        this.purchasedLotto = new PurchasedLotto(purchaseLotto());
-    }
-
-    public List<Lotto> getLotto() {
-        return Collections.unmodifiableList(purchasedLotto.getPurchasedLotto());
-    }
-
-    public List<List<Integer>> getLottoNumbers() {
-        return purchasedLotto.getLottoNumbers();
-    }
-
-    private List<Lotto> purchaseLotto() {
-        List<Lotto> lotto = new ArrayList<>();
-        for (int i = 0; i < lottoAmount.getPurchasedLottoCount(); i++) {
-            lotto.add(new Lotto(LOTTO_NUMBER_STRATEGY.randomNumbers()));
-        }
-        return lotto;
     }
 }
